Figured out baseline of our camera system.
Camera can now ease in and out and I have figured out the general Math behind it so I can tweak as needed. A few more kinks to work out but I may be able to work around them.

KNOWN ISSUE. If camera is supposed to rotate left, but say you are at 20deg and you want to go to 350 or -10deg. It will rotate all the way around to the right to get to it. Brainstorming possible fixes atm.